update to evp

+++ src/cpp/dns_cache.cpp
@@ -328,7 +328,7 @@ void dns_entry::initiate_lookup(const char *host, int port, const std::function<
     dns_map.erase(dns_map.find(host));
     throw std::system_error(rslt, std::system_category());
   }
-  auto sz = std::max(64 * 1024, PTHREAD_STACK_MIN);
+  auto sz = std::max(64 * 1024, (int)PTHREAD_STACK_MIN);
   rslt = pthread_attr_setstacksize(&nc->ptattr_, sz);
   if (rslt != 0)
   {
